Remote controlled preprocessor rules (#586)
