+++ proxy/src/main/java/com/wavefront/agent/preprocessor/PreprocessorUtil.java
@@ -1,21 +1,7 @@
 package com.wavefront.agent.preprocessor;
 
 import javax.annotation.Nullable;
-import java.util.Calendar;
 import java.util.Map;
-import java.util.TimeZone;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
-
-import com.mdimension.jchronic.Chronic;
-import com.mdimension.jchronic.Options;
-
-import wavefront.report.Annotation;
-import wavefront.report.ReportPoint;
-import wavefront.report.Span;
-
-import static com.wavefront.ingester.AbstractIngesterFormatter.unquote;
-import static org.apache.commons.lang3.ObjectUtils.firstNonNull;
 
 /**
  * Utility class for methods used by preprocessors.
